Move evaluation.py to higher module
